Don't use CFR mode if forced framerate is same as the source file

+++ src/avisynth/avssources.h
@@ -45,8 +45,8 @@ class AvisynthVideoSource : public IClip {
     VideoInfo VI;
     bool HighBitDepth;
     FFMS_VideoSource *V;
-    int FPSNum;
-    int FPSDen;
+    int64_t FPSNum;
+    int64_t FPSDen;
     int RFFMode;
     std::vector<FrameFields> FieldList;
     const char *VarPrefix;
